+++ lib/ui/settings/widgets/settings_contacts.dart
@@ -1,16 +1,23 @@
 import 'package:flutter/material.dart';
 
-import '../../../l10n/generated/app_localizations.dart';
 import '../../core/themes/dimensions.dart';
 import '../../core/widgets/widgets.dart';
 
-class SettingsContacts extends StatelessWidget {
-  const SettingsContacts({super.key});
+class SettingsAction extends StatelessWidget {
+  const SettingsAction({
+    super.key,
+    required this.title,
+    required this.subtitle,
+    required this.action,
+  });
+
+  final String title;
+  final String subtitle;
+  final CustomIconButton action;
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    final localizations = AppLocalizations.of(context)!;
     return Padding(
       padding: const EdgeInsets.all(AppDimensions.paddingMedium),
       child: Column(
@@ -19,23 +26,11 @@ class SettingsContacts extends StatelessWidget {
         children: [
           Row(
             children: [
-              Expanded(
-                child: Text(
-                  localizations.settingsContactTitle,
-                  style: theme.textTheme.titleMedium,
-                ),
-              ),
-              CustomIconButton.redirectExternal(
-                icon: CustomIcons.sendMail,
-                url: 'mailto:nosebitestudios@gmail.com',
-                tooltip: localizations.settingsContactEmailLabel,
-              ),
+              Expanded(child: Text(title, style: theme.textTheme.titleMedium)),
+              action,
             ],
           ),
-          Text(
-            localizations.settingsContactSubtitle,
-            style: theme.textTheme.bodyMedium,
-          ),
+          Text(subtitle, style: theme.textTheme.bodyMedium),
         ],
       ),
     );