+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.0-rc.2] - 2026-04-13
+## [1.0.0-rc.3] - 2026-04-14
 
 ### Added
+
+- Added a contact section to the settings page with email address for user feedback and contributions.
+- Added the ability to remove old articles from the app to free up storage space.
+
+## [1.0.0-rc.2] - 2026-04-13
+
+### Fixed
 
 - Fix AndroidManifest.xml to include INTERNET permission for network access.